PIRHDMI v1.11
removed dash causing unhandled exception... Added variables to
config.example.js file (config.pirhdmi.Debug and config.pirhdmi.Enable)
if config.pirhdmi is not in config.js this will default to disabled.

+++ main.js
@@ -95,28 +95,27 @@ kwsProcess.stdout.on('data', function (data) {
     console.log(data.toString())
 })
 
-// Get PIR-HDMI config
-if(typeof config.pir-hdmi == 'undefined'){
-  config.pir-hdmi = {}
+// Get PIRHDMI config
+if(typeof config.pirhdmi == 'undefined'){
+  config.pirhdmi = {}
 }
-var pirPin = config.pir-hdmi.PirPin || 26
-var ScreenTimeOut = config.pir-hdmi.ScreenTimeOut || 5.0
+var pirPin = config.pirhdmi.PirPin || 26
+var pirDebug = config.pirhdmi.Debug || true
+var ScreenTimeOut = config.pirhdmi.ScreenTimeOut || 0.5
+var pirEnable = config.pirhdmi.Enable || false
 
 // Initilize the PIR-HDMI process
-var PirProcess = spawn('python', ['./pir-hdmi/pir-hdmi.py', pirPin, ScreenTimeOut], {detached: false})
-// Handel messages from python script
+if (pirEnable == true){
+var PirProcess = spawn('python', ['./pir-hdmi/pir-hdmi.py', pirPin, ScreenTimeOut, pirDebug], {detached: false})
+// Handle messages from python script
 PirProcess.stderr.on('data', function (data) {
     var message = data.toString()
-    if(message.startsWith('INFO')){
-        // When a keyword is spotted, ping the speech service
-        mainWindow.webContents.send('PIR-Triggered', true)
-    }else{
-        console.error(message)
-    }
+    console.error(message)
 })
 PirProcess.stdout.on('data', function (data) {
-    console.log(data.toString())
+    console.error(data.toString())
 })
+}
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
@@ -130,5 +129,7 @@ app.on('window-all-closed', function () {
 // No matter how the app is quit, we should clean up after ourselvs
 app.on('will-quit', function () {
   kwsProcess.kill()
-  PirProcess.kill()
+  if (pirEnable==true){
+	  PirProcess.kill()
+  }
 })
